+++ NEWS.md
@@ -1,4 +1,7 @@
 - Additional methods for pre-allocated result arrays and `*Config` instances have been added to the ForwardDiff extension. [#871].
+MixedModels v5.3.0 Release Notes
+==============================
+- Implement `sparseL` as a specialization of `sparsemat`. Replace `_coord` utility with `_findnz` which, in most cases, falls through to `SparseArrays.findnz`. [#880]
 
 MixedModels v5.2.2 Release Notes
 ==============================
@@ -728,3 +731,4 @@ Package dependencies
 [#873]: https://github.com/JuliaStats/MixedModels.jl/issues/873
 [#875]: https://github.com/JuliaStats/MixedModels.jl/issues/875
 [#876]: https://github.com/JuliaStats/MixedModels.jl/issues/876
+[#880]: https://github.com/JuliaStats/MixedModels.jl/issues/880